requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
httpx==0.25.2
numpy==1.24.3
geopy==2.4.0
python-multipart==0.0.6
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
redis==5.0.1
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
celery==5.3.4

# Additional dependencies for ROS Pipeline
requests==2.31.0
aiofiles==23.2.1
python-json-logger==2.0.7
